Move `parse_animation` to the `animation` module
